completion: rename --track option of __git_complete_refs
The __git_complete_refs uses the "--track" option to specify when to
enable listing of unique remote branches which could be used by the
DWIMery of git checkout and git switch.

This is confusing, because both git checkout and git switch have
a --track option. This makes looking at the completion code for these
two functions confusing.

Rename this parameter to --dwim. Because it is plausible for users to
have developed their own completions which rely on __git_complete_ref,
keep --track as a synonym for --dwim, but do not use it in our own
completion code.

This reduces the confusion when seeing --track in the _git_switch() and
_git_checkout() completion functions.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>